+++ src/utils.c
@@ -39,7 +39,7 @@ SEXP rnng_scm(void) {
 
 /* ncurl - minimalist http client ------------------------------------------- */
 
-SEXP rnng_ncurl(SEXP http, SEXP args) {
+SEXP rnng_ncurl(SEXP http, SEXP method, SEXP ctype, SEXP auth, SEXP data) {
 
   nng_url *url;
   nng_http_client *client;
@@ -48,7 +48,7 @@ SEXP rnng_ncurl(SEXP http, SEXP args) {
   nng_aio *aio;
   int xc;
   uint16_t code;
-  void *data;
+  void *dat;
   size_t sz;
   struct nng_tls_config *cfg;
   int tls = 0;
@@ -68,15 +68,41 @@ SEXP rnng_ncurl(SEXP http, SEXP args) {
     nng_url_free(url);
     return Rf_ScalarInteger(xc);
   }
-  if (args != R_NilValue) {
-    const char *method = CHAR(STRING_ELT(VECTOR_ELT(args, 0), 0));
-    const char *ctype = CHAR(STRING_ELT(VECTOR_ELT(args, 1), 0));
-    const SEXP data = VECTOR_ELT(args, 2);
+  if (method != R_NilValue) {
+    const char *met = CHAR(STRING_ELT(method, 0));
+    xc = nng_http_req_set_method(req, met);
+    if (xc) {
+      nng_http_req_free(req);
+      nng_http_client_free(client);
+      nng_url_free(url);
+      return Rf_ScalarInteger(xc);
+    }
+  }
+  if (ctype != R_NilValue) {
+    const char *cty = CHAR(STRING_ELT(ctype, 0));
+    xc = nng_http_req_set_header(req, "Content-Type", cty);
+    if (xc) {
+      nng_http_req_free(req);
+      nng_http_client_free(client);
+      nng_url_free(url);
+      return Rf_ScalarInteger(xc);
+    }
+  }
+  if (auth != R_NilValue) {
+    const char *aut = CHAR(STRING_ELT(auth, 0));
+    xc = nng_http_req_set_header(req, "Authorization", aut);
+    if (xc) {
+      nng_http_req_free(req);
+      nng_http_client_free(client);
+      nng_url_free(url);
+      return Rf_ScalarInteger(xc);
+    }
+  }
+  if (data != R_NilValue) {
     unsigned char *dp = RAW(data);
     const R_xlen_t dlen = XLENGTH(data) - 1;
-    if ((xc = nng_http_req_set_method(req, method)) ||
-        (xc = nng_http_req_set_header(req, "Content-Type", ctype)) ||
-        (xc = nng_http_req_set_data(req, dp, dlen))) {
+    xc = nng_http_req_set_data(req, dp, dlen);
+    if (xc) {
       nng_http_req_free(req);
       nng_http_client_free(client);
       nng_url_free(url);
@@ -152,10 +178,10 @@ SEXP rnng_ncurl(SEXP http, SEXP args) {
     return ret;
   }
 
-  nng_http_res_get_data(res, &data, &sz);
+  nng_http_res_get_data(res, &dat, &sz);
   SEXP vec = PROTECT(Rf_allocVector(RAWSXP, sz));
   unsigned char *rp = RAW(vec);
-  memcpy(rp, data, sz);
+  memcpy(rp, dat, sz);
 
   if (tls)
     nng_tls_config_free(cfg);